refactor: Move initial file copy to non-blocking scene
Previously, copying files from the PDX bundle to the data directory on first launch was a synchronous, blocking operation. If a large number of ROMs or other assets were included, this could take several seconds, causing the Playdate OS to terminate the app for unresponsiveness.

This change introduces a new ⁠FileCopyingScene that handles this process asynchronously. The scene now performs the following steps:
- Collects a list of all files that need to be copied from the PDX. - Copies one file per frame, which prevents the main thread from blocking. - Displays a progress indicator to the user, providing feedback that the app has not stalled. - Once all files are copied, it transitions to the game scanning scene.

This resolves the crash issue on startup and significantly improves the user experience during the initial setup process. Closes #42

Signed-off-by: Toni Förster <toni.foerster@icloud.com>