fix(update): Defer update acknowledgment until notification is shown
Corrects a logic issue where a new update could be permanently missed by the user.

Previously, the new version number was written to disk as "seen" immediately upon network detection. If the user was in a game, the notification was correctly deferred, but if they quit the app without returning to the library, they would never be notified of that update again.

This change decouples the detection of an update from the acknowledgment of its notification:

- The network callback now only reports that a new version is available.
- A new function, `version_update_notification_shown()`, is introduced to handle writing the "seen" status to disk.
- This function is now called exclusively by the UI layer at the exact moment the update modal is presented to the user.

This ensures that the update is only marked as seen after the user has actually been notified, guaranteeing that pending update notifications are reliably displayed.

Signed-off-by: Toni Förster <toni.foerster@icloud.com>