feat(settings): Add preference to skip CGB confirmation
This adds a new preference, `skip_cgb_confirm`, to allow users to bypass the confirmation dialog for games flagged as CGB-only.

When this option is enabled, the game scene will automatically proceed with booting the game, setting the `ignore_cgb_check` flag, instead of displaying the warning message and waiting for user input.

The toggle for this preference has been added to the "Miscellaneous" section of the settings scene.

Signed-off-by: Toni Förster <toni.foerster@icloud.com>